refactor(pgpm): unify module emitter in @pgpmjs/transform (writeModule + PgpmModuleModel)
Phase 1 of the dials emission cleanup. The engine (semantic diff, granularity,
partition) already produces plan-ordered PgpmRows; this promotes the on-disk
module writer into @pgpmjs/transform as a single seam so pgpm diff, transform,
and import all render through one writer instead of each carrying its own copy.

- add PgpmModuleModel { name, requires, rows } + writeModule/writeControlFile/
  checkOverwrite in transform/src/module-emit.ts (imports the low-level plan/sql
  writers from @pgpmjs/ast to avoid a core->transform cycle)
- cli/utils/emit-package.ts becomes a thin re-export (writePackage=writeModule,
  EmitPackage=PgpmModuleModel) so transform/import commands are untouched
- pgpm diff drops its bespoke writeMigrationModule/writeControlFile and emits
  through the shared writeModule
