+++ pgpm/cli/src/commands/diff.ts
@@ -1,4 +1,4 @@
-import { PgpmMigrate, PgpmRow, SqlWriteOptions, writePgpmFiles, writePgpmPlan } from '@pgpmjs/core';
+import { PgpmMigrate } from '@pgpmjs/core';
 import {
   deltaChangesToRows,
   DiffSide,
@@ -16,7 +16,8 @@ import {
   SemanticDiffResult,
   SemanticObjectDiff,
   snapshotCatalog,
-  withoutColumnOrder
+  withoutColumnOrder,
+  writeModule
 } from '@pgpmjs/transform';
 import { spawn } from 'child_process';
 import * as fs from 'fs';
@@ -121,36 +122,6 @@ const loadSide = async (spec: string, cwd: string): Promise<DiffSide> => {
 /** `resolveDiffSideKind` without throwing on relative on-disk paths. */
 const resolveDiffSideKindSafe = (spec: string): 'database' | 'disk' =>
   /^postgres(ql)?:\/\//.test(spec) || spec.startsWith('db:') ? 'database' : 'disk';
-
-const writeControlFile = (dir: string, name: string): void => {
-  const lines = [
-    `# ${name} extension`,
-    `comment = '${name} migration (generated by pgpm diff)'`,
-    `default_version = '0.0.1'`,
-    `relocatable = false`,
-    `superuser = false`
-  ];
-  fs.writeFileSync(path.join(dir, `${name}.control`), lines.join('\n') + '\n');
-};
-
-/** Write the delta as a pgpm module at <outBase>/<pkgName>. */
-const writeMigrationModule = (outBase: string, pkgName: string, rows: PgpmRow[]): string => {
-  const dir = path.join(outBase, pkgName);
-  fs.rmSync(path.join(dir, 'deploy'), { recursive: true, force: true });
-  fs.rmSync(path.join(dir, 'revert'), { recursive: true, force: true });
-  fs.rmSync(path.join(dir, 'verify'), { recursive: true, force: true });
-  fs.mkdirSync(dir, { recursive: true });
-
-  writeControlFile(dir, pkgName);
-  const opts: SqlWriteOptions = {
-    outdir: outBase,
-    name: pkgName,
-    replacer: (str: string) => str.split('constructive-extension-name').join(pkgName)
-  };
-  writePgpmPlan(rows, opts);
-  writePgpmFiles(rows, opts);
-  return dir;
-};
 
 const createScratchDb = async (config: PgConfig, dbName: string): Promise<void> => {
   const adminPool = getPgPool({ ...config, database: 'postgres' });
@@ -330,7 +301,7 @@ export default async (
     if (rows.length === 0) {
       log.info('no migration changes to emit (sides are identical).');
     } else {
-      migrationDir = writeMigrationModule(outBase, pkgName, rows);
+      migrationDir = writeModule(outBase, { name: pkgName, requires: [], rows });
       if (emitMigration) log.success(`wrote ${rows.length} migration change(s) to ${migrationDir}`);
     }
   }
